Enable Android module and upgrade Gradle
Re-enables the Android app module and restores Android build tooling to support building the Android target.

Adds Kotlin Android and Android application plugins to the build script and bumps the Gradle wrapper to 8.13 to match AGP requirements. This resolves previous AGP dependency issues and allows the Android module to be included in the multi-module build.